src/requirements.txt
openai==1.36.0
python-dotenv==1.0.1
